Move `orm_utils.py` to `tests/lib`
